\name{converge.gauss}
\alias{converge.gauss}
\title{Test if a gaussian model fit has converged}
\usage{
  \method{converge}{gauss} (object, ...)
}
\arguments{
  \item{object}{a class of model fit}

  \item{...}{additional arguments}
}
\value{
  convergence status (logical)
}
\description{
  Test if a gaussian model fit has converged
}

