\name{fitted.gauss}
\alias{fitted.gauss}
\title{extract the fitted model parameters}
\usage{
  \method{fitted}{gauss} (object, ...)
}
\arguments{
  \item{object}{a gauss class object}

  \item{...}{aditional arguments (currently ignored)}
}
\value{
  a (named) vector of the MLE parameters
}
\description{
  extract the fitted model parameters
}

