\name{logLik.gauss}
\alias{logLik.gauss}
\title{extract the logLik}
\usage{
  \method{logLik}{gauss} (object, ...)
}
\arguments{
  \item{object}{a gauss class model fit}
}
\value{
  the log likelihood (not the negative log lik, & not 2*log
  likelihood)
}
\description{
  extract the logLik
}
\details{
  note that this is an extraction method, not a calculation
  method
}

