\name{simulate.gauss}
\alias{simulate.gauss}
\title{simulate method}
\usage{
  \method{simulate}{gauss} (object, nsim = 1, seed = NULL,
    ...)
}
\arguments{
  \item{object}{any gauss-class object}

  \item{...}{additional arguments (currently ignored)}
}
\value{
  a data frame with the time simulated values
}
\description{
  simulate method
}

